requirements.txt
python-docx==0.8.11
beautifulsoup4==4.9.0
htmldocx==0.0.6
